requirements.txt
flask==2.3.3
flask-cors==4.0.0
Werkzeug==2.3.7
python-dotenv==1.0.0
mysql-connector-python==8.0.33
PyJWT==2.8.0
google-cloud-language==2.13.0
pandas==2.1.0
matplotlib==3.8.0
scikit-learn==1.3.0
numpy==1.24.3
openai==1.3.0
pytest==7.4.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 